Fix minikube based on reviews
