+++ tests/util/localregistry/cmd/main.go
@@ -25,4 +25,3 @@ func main() {
 		os.Exit(-1)
 	}
 }
-
